+++ CHANGELOG.md
@@ -11,6 +11,7 @@ GAMSPy 1.6.0
   - Allow registering metadata to symbols via `symbol._metadata` attribute.
   - Allow indexing into symbols with integers.
   - Add `bypass_solver` option.
+  - Add configurable options via set_options and get_option.
 - Documentation
   - Add CNNs to embedding Neural Network documentation.
 
@@ -22,7 +23,6 @@ GAMSPy 1.5.1
   - Fix debugging level bug of NEOS backend.
   - Fix license issue of big models that are solved with frozen solve.
   - Allow loadRecordsFromGdx to domain forward.
-  - Upgrade gamspy_base and gamsapi dependencies.
   - Enhance bound propagation for `MaxPool2d` and `MinPool2d` classes.
 - Testing
   - Add bound propagation tests for `MaxPool2d` and `MinPool2d` classes.